feat: Merge WeatherApp into CommandPanel
This commit merges the `WeatherApp` GUI component directly into the `butler/CommandPanel.py` file.

This change was made to follow the user's request to "move" the weather app into the command panel, resulting in a more tightly integrated component.

Changes include:
- The `WeatherApp` class and its dependencies were moved from the separate `main.py` file into `butler/CommandPanel.py`.
- The `CommandPanel` now launches the `WeatherApp` directly, instead of using the generic program execution mechanism.
- The now-redundant `main.py` file has been deleted.
- The entry for the weather app in `butler/program_mapping.json` has been removed as it is no longer needed.
